+++ tests/database.js
@@ -104,7 +104,7 @@ test('database: get tags', t => {
       { name: 'namespace:e', fileCount: 1 },
       { name: 'namespace:d', fileCount: 1 },
       { name: 'namespace:c', fileCount: 1 },
-      { name: 'namespace:b', fileCount: 1 },
+      { name: 'namespace:b', fileCount: 1 }
     ]
   )
 })
@@ -976,6 +976,64 @@ test('database: get files by tags', t => {
         '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
       thumbnailUrl: thumbnailsBaseUrl +
         '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+    }]
+  )
+})
+
+test('database: get files by excluded tags', t => {
+  t.deepEqual(
+    files.getByTags(1, ['-sit', '-amet']),
+    [
+      {
+        id: 5,
+        mime: 'image/png',
+        size: 6672,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/d2f5788f623cde1f0fb3dc801396fee235c67ed11d9452bfd765f1331587401d',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/d2f5788f623cde1f0fb3dc801396fee235c67ed11d9452bfd765f1331587401d'
+      },
+      {
+        id: 4,
+        mime: 'image/png',
+        size: 6665,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      }
+    ]
+  )
+})
+
+test('database: get files by tags and excluded tags', t => {
+  t.deepEqual(
+    files.getByTags(1, ['dolor', '-sit', '-amet']),
+    [{
+      id: 3,
+      mime: 'image/png',
+      size: 6117,
+      width: 500,
+      height: 500,
+      mediaUrl: originalBaseUrl +
+        '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+      thumbnailUrl: thumbnailsBaseUrl +
+        '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
     }]
   )
 })